+++ packages/core/src/lib/date.js
@@ -1,4 +1,4 @@
-import {isDate} from './utils.js';
+import {isDate, isFunction} from './utils.js';
 
 export const DAY_IN_SECONDS = 86400;
 
@@ -180,6 +180,16 @@ export function getWeekNumber(date, firstDay) {
     let yearStart = new Date(Date.UTC(date.getUTCFullYear(),0,1));
     // Calculate full weeks to `date`
     return Math.ceil((((date - yearStart) / 1000 / DAY_IN_SECONDS) + 1) / 7);
+}
+
+export function createWeekNumberContent(week, weekNumberContent, date) {
+    if (weekNumberContent) {
+        return isFunction(weekNumberContent)
+            ? weekNumberContent({date: toLocalDate(date), week})
+            : weekNumberContent;
+    }
+
+    return 'W' + String(week).padStart(2, '0');
 }
 
 /**